更新IDOBluetooth 3.1.4版本
1、优化锐捷手环连接服务特征多次回调bug;
2、添加运动模式排序demo

+++ IDOBlue/IDOBlue/Models/ViewModels/SetViewModel/SetSportShortcutViewModel.m
@@ -74,7 +74,7 @@
         NSIndexPath * indexPath = [funcVC.tableView indexPathForCell:tableViewCell];
         SwitchCellModel * switchCellModel = [strongSelf.cellModels objectAtIndex:indexPath.row];
         if (strongSelf.selectedCount >= __IDO_FUNCTABLE__.sportShowCount && onSwitch.isOn) {
-            [funcVC showToastWithText:[NSString stringWithFormat:@"最多支持%ld种类型",__IDO_FUNCTABLE__.sportShowCount]];
+            [funcVC showToastWithText:[NSString stringWithFormat:@"最多支持%ld种类型",(long)__IDO_FUNCTABLE__.sportShowCount]];
             switchCellModel.data = @[@(NO)];
             [funcVC.tableView reloadRowsAtIndexPaths:@[indexPath] withRowAnimation:UITableViewRowAnimationNone];
             return ;
